+++ src/internal/interpreter/map-interpreter.errors.test.ts
@@ -1,4 +1,5 @@
 import { MapASTNode, MapDocumentNode, MapHeaderNode } from '@superfaceai/ast';
+import { parseMap, Source } from '@superfaceai/parser';
 import { getLocal } from 'mockttp';
 
 import { CrossFetch } from '../../lib/fetch';
@@ -25,6 +26,16 @@ const header: MapHeaderNode = {
   },
   provider: 'example',
 };
+
+const parseMapFromSource = (source: string) =>
+  parseMap(
+    new Source(
+      `
+      profile = "example@0.0"
+      provider = "example"
+      ` + source
+    )
+  );
 
 describe('MapInterpreter errors', () => {
   describe('MapASTError', () => {
@@ -764,6 +775,33 @@ AST Path: definitions[0].statements[0].assignments[0].value`
       const result = await interpreter.perform(ast);
       expect(result.isErr() && result.error instanceof JessieError).toEqual(
         true
+      );
+    });
+
+    it('should return error when using parameters in baseUrl, but they are not supplied', async () => {
+      const url = '/twelve/something';
+      await mockServer.get(url).thenJson(200, { data: 12 });
+      const ast = parseMapFromSource(`
+        map Test {
+          http GET "/something" {
+            response 200 "application/json" {
+              map result {
+                result = body.data
+              }
+            }
+          }
+        }`);
+      const interpreter = new MapInterpreter(
+        {
+          usecase: 'Test',
+          security: [],
+          serviceBaseUrl: `${serviceBaseUrl}/{path}`,
+        },
+        { fetchInstance }
+      );
+      const result = await interpreter.perform(ast);
+      expect(result.isErr() && result.error.toString()).toMatch(
+        'Missing values for URL path replacement: path'
       );
     });
   });